refactor: reorganize RV32IM frontend ownership
Move RV32IM public API exports into api modules and split the old flat chunk files into chunk/fold, chunk/step_ivc, chunk/step_relation, and chunk/transition so the per-chunk proof flow is readable by ownership.

Split RV32IM Construction-2 and F' code into construction2 and f_prime modules, including the fresh-instance builder surface, F' accumulator, and side-lane authority code.

Move IVC SNARK and terminal F' committed code under ivc_snark, split kernel/simple into context/support/types, and give kernel openings their own mod and claim-reduction error surface.

Break the main_relation_spartan machinery into chunk_diagnostics, chunk_step, recursive_step, and focused diagnostic files so the recursive proof path no longer lives in multi-thousand-line files.

Move the RV32IM recursion shape probe implementation into support code and update RV32IM/f_prime_conformance tests to follow the new module paths without changing the optimized folding mode behavior.

+++ crates/neo-fold-next/tests/rv32im_parity.rs
@@ -228,12 +228,12 @@ fn parity_cases_reach_expected_halted_states() {
                         && row.rd >= INLINE_SCRATCH_REGISTER_BASE
                 }));
                 assert_eq!(derived.kernel.final_pc, 16);
-                assert_eq!(derived.kernel.final_registers[1], 1 << 40);
-                assert_eq!(derived.kernel.final_registers[2], 1 << 35);
-                assert_eq!(derived.kernel.final_registers[3], 1 << 41);
-                assert_eq!(derived.kernel.final_registers[4], 1 << 34);
-                assert_eq!(derived.kernel.final_registers[5], 1 << 42);
-                assert_eq!(derived.kernel.final_registers[6], 1 << 33);
+                assert_eq!(derived.kernel.final_registers[1], (1u64 << 40) as u32);
+                assert_eq!(derived.kernel.final_registers[2], (1u64 << 35) as u32);
+                assert_eq!(derived.kernel.final_registers[3], (1u64 << 41) as u32);
+                assert_eq!(derived.kernel.final_registers[4], (1u64 << 34) as u32);
+                assert_eq!(derived.kernel.final_registers[5], (1u64 << 42) as u32);
+                assert_eq!(derived.kernel.final_registers[6], (1u64 << 33) as u32);
                 assert_eq!(derived.kernel.final_registers[7], 2048);
                 assert_eq!(derived.kernel.final_registers[8], 2048);
                 assert_eq!(derived.kernel.final_registers[9], 2048);
